+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Added `\Wizaplace\SDK\Vendor\Order\OrderService::createShipment`
 - Added `\Wizaplace\SDK\Vendor\Order\OrderService::setInvoiceNumber`
 - Added `\Wizaplace\SDK\User\User::getCompanyId`
+- Added `\Wizaplace\SDK\User\User::isVendor`
 
 ### Bugfixes